Switch to mixin
